fix:> teamPOST 오류 수정 #61
